rename: MySetting관련 컴포넌트 폴더 이동, 파일명 및 확장자 수정
