+++ src/components/MyPage/MySetting/EditInfo/EditNicknameInfo/index.tsx
